HTML/JS/CSS Exercise - HTML Table Generator
HTML exercise for a potential employer.

Instructions:
Q3: Given an input JSON collection...
employees: [
    { firstName: ..., lastName: ..., age: ... }
]
... dynamically populate an HTML table with employee names columnwise.

i.e. top to bottom then left to right.
Highlight every 4th name in red text.
The number of rows in the table are fixed as an input variable 1 <=
nRows <= 100.
The number of columns are dynamically calculated to accommodate all
employees to be displayed.
For example, if nRows = 30 and there are 100 employees in the
collection, display 1 to 30 in column 1, 31 to 60 in column 2, 61 to 90
in column 3 and 91 onwards in column 4.

In certain cases, not all cells will be populated i.e. if (employee
count) < (rows x columns)

Make sure the names are sorted alphabetically on lastName (descending).

If two employees have the same lastName, sort on firstName (ascending).

Note:
A] This must be built using HTML, JS, CSS only.
You cannot use any server side constructs.
B] You may optionally use jQuery (http://jquery.com/) or Handlebars
(http://handlebarsjs.com/)
No other JS or CSS libraries may be used.